personality.py
personality.py is intended to be an AI with moods. I tried multiple different methods (The 2 closest to working are left in the hash comments as proof of hard work, and possible alternatives i was attempting), however I was ultimately unsuccessful.
The closest I came to success was getting output asking for a command, receiving it, and the AI only responding with the 'fear' emotion's output.
As well as this, I got one of the loops to send the output asking for a command, but before anything could be input, the line was immediately followed by the 'angry' emotion output message.
Neither of the above worked, but clearly show that I can adjust this program and fix it during free time in the future.