tennis player & simulation .py
I tried to get the entire tennis event to work in the single 'tennismatch.py' program, as I believed that would be an easier and a quicker/more optimal solution (rather than needing multiple programs). It did not work out completely, but my mindset is clearly described in the hash comments. I just wanted to submit what else was asked of us, so I added these two in order to have all three in the repository.