calc/project 4
this is the calc 4 project, submitted more for the fifth time as a re-push. i have hashed errors and possible solutions, as well as hashing ways we went over going about the code in class- and how i attempted to do what was asked of us. additionally, i have specifically addressed every bullet point on the project instructions.